requirements.txt
django
djangorestframework
django-cors-headers
pillow
pandas